backend/requirements.txt
fastapi
uvicorn
python-multipart
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
PyPDF2
pdfminer.six
requests
firebase-admin